refact : reafact canvas constructure
